refactor: extract refactor source planning helpers
